refactor: extract onboarding feature folder
Move splash-screen, welcome + tutor screens, doc-loader (renamed
load-or-create-doc-from-url), and welcome-content + random-writer-name
helpers (pulled from schema) into src/app/features/onboarding/.
