+++ src/app/features/onboarding/index.ts
@@ -1,0 +1,13 @@
+export {
+	SplashScreen,
+	SplashScreenStatic,
+	useSplashDelay,
+} from "./widgets/splash-screen"
+export { welcomeLoader } from "./screens/welcome-screen"
+export { tutorLoader } from "./screens/tutor-screen"
+export { loadOrCreateDocFromUrl } from "./lib/load-or-create-doc-from-url"
+export { getRandomWriterName } from "./lib/random-writer-name"
+export {
+	fetchWelcomeContent,
+	getSpaceWelcomeContent,
+} from "./lib/welcome-content"
