+++ src/app/features/onboarding/widgets/splash-screen.tsx
