+++ src/app/routes/welcome.tsx
@@ -1,8 +1,8 @@
 import { createFileRoute } from "@tanstack/react-router"
-import { loadOrCreateDoc } from "@/lib/doc-loader"
+import { welcomeLoader } from "@/app/features/onboarding"
 
 export { Route }
 
 let Route = createFileRoute("/welcome")({
-	loader: ({ context }) => loadOrCreateDoc(context.me, "/docs/welcome.md"),
+	loader: ({ context }) => welcomeLoader(context.me),
 })
